requirements.txt
streamlit
cassandra-driver
pymongo